+++ .github/workflows/build-wheels.yml
@@ -37,4 +37,4 @@ jobs:
 
       - name: Install the Python wheel
         run: |
-          python3 -m pip install dist/aleph_client-*.whl
+          python3 -m pip install dist/aleph_sdk_python-*.whl
