+++ docker/python-3.9.dockerfile
@@ -7,9 +7,9 @@ RUN apt-get update && apt-get -y upgrade && apt-get install -y \
 
 RUN useradd -s /bin/bash --create-home user
 RUN mkdir /opt/venv
-RUN mkdir /opt/aleph-client/
+RUN mkdir /opt/aleph-sdk-python/
 RUN chown user:user /opt/venv
-RUN chown user:user /opt/aleph-client
+RUN chown user:user /opt/aleph-sdk-python
 
 USER user
 RUN python3 -m venv /opt/venv
@@ -23,10 +23,10 @@ RUN pip install --upgrade secp256k1 coincurve aiohttp eciespy python-magic typer
 RUN pip install --upgrade 'aleph-message~=0.3.0' eth_account pynacl base58
 RUN pip install --upgrade pytest pytest-cov pytest-asyncio mypy types-setuptools pytest-asyncio fastapi httpx requests
 
-WORKDIR /opt/aleph-client/
+WORKDIR /opt/aleph-sdk-python/
 COPY . .
 USER root
-RUN chown -R user:user /opt/aleph-client
+RUN chown -R user:user /opt/aleph-sdk-python
 
 RUN pip install -e .[testing,ethereum,solana,tezos]
