+++ docker/with-ipfs.dockerfile
@@ -25,8 +25,8 @@ RUN apt-get update && apt-get -y upgrade && apt-get install -y \
      libsecp256k1-dev \
      && rm -rf /var/lib/apt/lists/*
 
-RUN mkdir /opt/aleph-client/
-WORKDIR /opt/aleph-client/
+RUN mkdir /opt/aleph-sdk-python/
+WORKDIR /opt/aleph-sdk-python/
 COPY . .
 
 RUN pip install -e .[testing,ethereum]
