+++ examples/mqtt.py
@@ -8,10 +8,10 @@ from typing import Dict
 import aiomqtt
 import click
 
-from aleph_client import AuthenticatedUserSession
-from aleph_client.chains.common import get_fallback_private_key
-from aleph_client.chains.ethereum import ETHAccount
-from aleph_client.conf import settings
+from aleph.sdk.chains.common import get_fallback_private_key
+from aleph.sdk.chains.ethereum import ETHAccount
+from aleph.sdk.conf import settings
+from aleph.sdk.user_session import AuthenticatedUserSession
 
 
 def get_input_data(value):
