+++ setup.cfg
@@ -3,14 +3,14 @@
 # http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
 
 [metadata]
-name = aleph-client
+name = aleph-sdk-python
 description = Lightweight Python Client library for the Aleph.im network
 author = Aleph.im Team
 author_email = hello@aleph.im
 license = mit
 long_description = file: README.rst
 long_description_content_type = text/x-rst; charset=UTF-8
-url = https://github.com/aleph-im/aleph-client
+url = https://github.com/aleph-im/aleph-sdk-python
 project_urls =
     Documentation = https://aleph.im/
 # Change if running only on Windows, Mac or Linux (comma-separated)
@@ -51,7 +51,7 @@ exclude =
 
 [options.extras_require]
 # Add here additional requirements for extra features, to install with:
-# `pip install aleph-client[PDF]` like:
+# `pip install aleph-sdk-python[PDF]` like:
 # PDF = ReportLab; RXP
 # Add here test requirements (semicolon/line-separated)
 testing =
