+++ tests/integration/itest_posts.py
@@ -1,7 +1,7 @@
 import pytest
 from aleph_message.models import MessagesResponse
 
-from aleph_client.user_session import AuthenticatedUserSession
+from aleph.sdk.user_session import AuthenticatedUserSession
 from tests.integration.toolkit import try_until
 
 from .config import REFERENCE_NODE, TARGET_NODE
