+++ tests/unit/test_chain_nuls1.py
@@ -1,7 +1,7 @@
 import pytest
 from coincurve.keys import PrivateKey
 
-from aleph_client.chains.nuls1 import NulsSignature
+from aleph.sdk.chains.nuls1 import NulsSignature
 
 SECRET = (
     b"\xc4\xfe\xe65\x96\x14\xb4:\r: \x05;\x12j\x9bJ"
