+++ tests/unit/test_chain_nuls1_compat.py
@@ -9,8 +9,8 @@ import pytest
 import secp256k1
 from coincurve.keys import PrivateKey
 
-from aleph_client.chains.common import get_fallback_private_key
-from aleph_client.chains.nuls1 import LOGGER, MESSAGE_TEMPLATE, NulsSignature, VarInt
+from aleph.sdk.chains.common import get_fallback_private_key
+from aleph.sdk.chains.nuls1 import LOGGER, MESSAGE_TEMPLATE, NulsSignature, VarInt
 
 SECRET = (
     b"\xc4\xfe\xe65\x96\x14\xb4:\r: \x05;\x12j\x9bJ"
