+++ tests/unit/test_init.py
@@ -1,4 +1,4 @@
-from aleph_client import __version__
+from aleph.sdk import __version__
 
 
 def test_version():
